Store heading anchor in custom data
This is slightly cleaner design (separating parsing from rendering) at the cost of more code.

Unfortunately, it is not really possible to hook into the parser without forking cmark-gfm (the extensibility related functions are all internal) so we still need to do this by AST transformations.

Also, the fact that we use custom block element means that the HTML output is split over multiple lines.